requirements.txt
fastapi
uvicorn
python-multipart
python-docx
odfpy
pytesseract
Pillow
reportlab
